Crate actix_connect

Actix connect - tcp connector service

Package feature

ssl - enables ssl support via openssl crate

rust-tls - enables ssl support via rustls crate

Modules

ssl
SSL Services

Structs

AsyncResolver
A handle for resolving DNS records.

Connect
Connect request

Connection

ResolveError
The error type for errors that get returned in the crate

Resolver
DNS Resolver Service

ResolverConfig
Configuration for the upstream nameservers to use for resolution

ResolverFactory
DNS Resolver Service factory

ResolverOpts
Configuration for the Resolver

TcpConnector
Tcp connector service

TcpConnectorFactory
Tcp connector service factory

Enums

ConnectError

Traits

Address
Connect request

Functions

default_connector
Create connector service with default parameters

default_connector_factory
Create connector service factory with default parameters

new_connector
Create tcp connector service

new_connector_factory
Create tcp connector service

read_system_conf

start_default_resolver

start_resolver